Move JsonResolver to separate file
