+++ Saving/SerializeThisAttribute.cs
@@ -12,25 +12,4 @@ namespace KarpikQuests.Saving
         public Version Version { get; set; }
         public bool IsReference { get; set; }
     }
-
-    class JsonResolver : ISerializer<IQuestAggregator>
-    {
-        public IQuestAggregator? Deserialize(string data)
-        {
-            return JsonConvert.DeserializeObject(data, new JsonSerializerSettings
-            {
-                TypeNameHandling = TypeNameHandling.All,
-
-            }) as IQuestAggregator;
-        }
-
-        public string Serialize(IQuestAggregator aggregator)
-        {
-            return JsonConvert.SerializeObject(aggregator, new JsonSerializerSettings
-            {
-                Formatting = Formatting.Indented,
-                TypeNameHandling = TypeNameHandling.All
-            });
-        }
-    }
 }
